libOpenCOR: reorganised our files.
